+++ mock/data/download.js
@@ -1,9 +1,10 @@
 module.exports = [{
-    url: '/files/:name',
-    method: 'post',
+    url: '/download/:filename',
+    method: 'get',
     response: {
         delay: 2000,
         headers: {
+            'Content-Type': 'text/plain',
             'Content-Disposition': 'attachment;filename=sample.txt;'
         },
         body: '/sample.txt'
