+++ mock/data/sample.js
@@ -1,6 +1,3 @@
-var faker = require('faker');
-var Mock = require('mockjs');
-
 module.exports = [{
     url: '/user/123',
     response: {
@@ -14,31 +11,31 @@ module.exports = [{
     url: '/user/:id',
     response: {
         body: {
-            id: faker.random.uuid(),
-            name: faker.name.findName(),
-            email: faker.internet.email()
+            id: 123,
+            name: 'Stephen',
+            email: 'xxx@gmail.com'
         }
     }
 }, {
     url: '/user/*.do',
     response: {
-        body: Mock.mock({
-            'data|20': [{
-                id: '@integer(0, 10000)',
-                name: '@name',
-                email: '@email'
-            }]
-        }).data
+        body: {
+            id: 123,
+            name: 'Stephen',
+            email: 'xxx@gmail.com'
+        }
     }
 }, {
     url: '/user/**/list',
     response: {
-        body: Mock.mock({
-            'data|20': [{
-                id: '@integer(0, 10000)',
-                name: '@name',
-                email: '@email'
-            }]
-        }).data
+        body: [{
+            id: 123, 
+            name: 'Stephen',
+            age: 30
+        }, {
+            id: 124, 
+            name: 'Ricky',
+            age: 20				
+        }]
     }
 }];
